services/downloader/requirements.txt
flask
bs4
pymongo
jsonschema
flask-cors
python-dateutil
gunicorn
cloudscraper
